Move simple gp model to model_wrappers
