+++ emukit/model_wrappers/__init__.py
@@ -4,3 +4,4 @@
 
 from .gpy_model_wrappers import GPyModelWrapper, GPyMultiOutputWrapper  # noqa: F401
 from .gpy_quadrature_wrappers import BaseGaussianProcessGPy, RBFGPy  # noqa: F401
+from .simple_gp_model import SimpleGaussianProcessModel  # noqa: F401
