+++ emukit/model_wrappers/simple_gp_model.py
